bfcodec: the key length is 64-bit, and the NSData overload was narrowing it
-cipherInit:keyLength: took an int length where the encoding is Q. That is a
real 32-against-64-bit size difference rather than a signedness one, and the
binary confirms it directly: the prologue does mov x19,x3, saving the whole
64-bit register, and forwards the whole of x2 to the key schedule.

The parameter, the file-static SetBlowfishKey it feeds, and the cast at the
NSData overload are all corrected. That cast had been narrowing an NSUInteger
length back down to int before passing it, so the declared type and the call
site were compensating for each other.

The keyIndex modulo in SetBlowfishKey is unaffected: keyIndex starts at zero
and only increments, so promoting it against an unsigned length is safe.

Five width reports remain, all signedness rather than size.

+++ Project/BFCodec.h
@@ -38,7 +38,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @param length The length of @p key in bytes; the key is cycled modulo this length.
  * @ghidraAddress 0x1534c
  */
-- (void)cipherInit:(const char *)key keyLength:(int)length;
+- (void)cipherInit:(const char *)key keyLength:(NSUInteger)length;
 
 /**
  * @brief Encipher a mutable buffer in place with Blowfish-CBC.
